+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- [SIL.Core] Added String extension method SplitLines.
+
 ## [15.0.0] - 2025-01-06
 
 ### Added